refactor in Transaction db
