+++ lib/HistoricSync.js
@@ -2,12 +2,9 @@
 
 var imports        = require('soop').imports();
 var util           = require('util');
-var assert         = require('assert');
 var async          = require('async');
 
 var bitcore        = require('bitcore');
-var RpcClient      = bitcore.RpcClient;
-var Script         = bitcore.Script;
 var networks       = bitcore.networks;
 var config         = imports.config || require('../config/config');
 var Sync           = require('./Sync');
@@ -15,23 +12,7 @@ var sockets        = require('../app/controllers/socket.js');
 var BlockExtractor = require('./BlockExtractor.js');
 var buffertools    = require('buffertools');
 var bitcoreUtil    = bitcore.util;
-var Bignum         = bitcore.Bignum;
-var Script       = bitcore.Script;
-// This is 0.1.2 = > c++ version of base58-native
- var base58       = require('base58-native').base58Check;
- var encodedData  = require('soop').load('bitcore/util/EncodedData',{
-   base58: base58
- });
- var versionedData= require('soop').load('bitcore/util/VersionedData',{
-   parent: encodedData
- });
-
-var Address = require('soop').load('bitcore/lib/Address',{
-  parent: versionedData
-});
-
 var logger = require('./logger').logger;
-var d      = logger.log;
 var info   = logger.info;
 var error  = logger.error;
 
@@ -140,48 +121,6 @@ HistoricSync.prototype.getBlockFromRPC = function(cb) {
   });
 };
 
-HistoricSync.prototype._fromBuffer = function (buf) {
-    var buf2 = buffertools.reverse(buf);
-    return parseInt(buf2.toString('hex'), 16);
-};
-
-HistoricSync.prototype.getStandardizedTx = function (tx, time, isCoinBase) {
-  var self = this;
-  tx.txid = bitcoreUtil.formatHashFull(tx.getHash());
-  var ti=0;
-
-  tx.vin = tx.ins.map(function(txin) {
-    var ret = {n: ti++};
-    if (isCoinBase) {
-      ret.isCoinBase = true;
-    } else {
-      ret.txid = buffertools.reverse(new Buffer(txin.getOutpointHash())).toString('hex');
-      ret.vout = txin.getOutpointIndex();
-    }
-    return ret;
-  });
-
-  var to = 0;
-  tx.vout = tx.outs.map(function(txout) {
-    var val;
-    if (txout.s) {
-      var s = new Script(txout.s);
-      var addrs = new Address.fromScriptPubKey(s, config.network);
-      // support only for p2pubkey p2pubkeyhash and p2sh
-      if (addrs && addrs.length === 1) {
-        val = {addresses: [addrs[0].toString() ] };
-      }
-    }
-    return {
-      valueSat: self._fromBuffer(txout.v),
-      scriptPubKey: val,
-      n: to++,
-    };
-  });
-  tx.time = time;
-  return tx;
-};
-
 HistoricSync.prototype.getStandardizedBlock = function(b) {
   var self = this;
 
@@ -192,7 +131,7 @@ HistoricSync.prototype.getStandardizedBlock = function(b) {
   };
   var isCoinBase = 1;
   block.tx = b.txs.map(function(tx){
-    var ret = self.getStandardizedTx(tx, b.timestamp, isCoinBase);
+    var ret = self.sync.txDb.getStandardizedTx(tx, b.timestamp, isCoinBase);
     isCoinBase=0;
     return ret;
   });
