+++ lib/PeerSync.js
@@ -49,10 +49,8 @@ PeerSync.prototype.handleInv = function(info) {
 
 PeerSync.prototype.handleTx = function(info) {
   var self =this;
-  var tx = info.message.tx.getStandardizedObject();
-  tx.outs = info.message.tx.outs;
-  tx.ins = info.message.tx.ins;
-  console.log('[p2p_sync] Handle tx: ' + tx.hash);
+  var tx = this.sync.txDb.getStandardizedTx(info.message.tx);
+  console.log('[p2p_sync] Handle tx: ' + tx.txid);
   tx.time = tx.time || Math.round(new Date().getTime() / 1000);
 
   this.sync.storeTxs([tx], function(err) {
@@ -62,9 +60,9 @@ PeerSync.prototype.handleTx = function(info) {
     else {
       if (self.shouldBroadcast) {
         sockets.broadcastTx(tx);
-        if (tx.addrsToEmit) {
-          for(var ii in tx.addrsToEmit){
-            sockets.broadcastAddressTx(tx.addrsToEmit[ii], tx.txid);
+        if (tx.relatedAddrs) {
+          for(var ii in tx.relatedAddrs){
+            sockets.broadcastAddressTx(tx.relatedAddrs[ii], tx.txid);
           }
         }
       }
@@ -76,9 +74,7 @@ PeerSync.prototype.handleTx = function(info) {
 PeerSync.prototype.handleBlock = function(info) {
   var self = this;
   var block = info.message.block;
-console.log('[PeerSync.js.78:block:]',block); //TODO
   var blockHash = bitcoreUtil.formatHashFull(block.calcHash());
-
   console.log('[p2p_sync] Handle block: %s (allowReorgs: %s)', blockHash, self.allowReorgs);
 
   var tx_hashes = block.txs.map(function(tx) {
