+++ lib/TransactionDb.js
@@ -28,15 +28,30 @@ var bitcore  = require('bitcore'),
     levelup  = require('levelup'),
     async    = require('async'),
     config   = require('../config/config'),
-    assert   = require('assert');
+    assert   = require('assert'),
+    Script          = bitcore.Script,
+    bitcoreUtil     = bitcore.util,
+    buffertools    = require('buffertools');
 
 var logger = require('./logger').logger;
-var d      = logger.log;
 var info   = logger.info;
-var warn   = logger.warn;
-var db           = imports.db || levelup(config.leveldb + '/txs',{maxOpenFiles: MAX_OPEN_FILES} );
 
+var db           = imports.db || levelup(config.leveldb + '/txs',{maxOpenFiles: MAX_OPEN_FILES} );
 var PoolMatch     = imports.poolMatch || require('soop').load('./PoolMatch',config);
+// This is 0.1.2 = > c++ version of base58-native
+ var base58       = require('base58-native').base58Check;
+ var encodedData  = require('soop').load('bitcore/util/EncodedData',{
+   base58: base58
+ });
+ var versionedData= require('soop').load('bitcore/util/VersionedData',{
+   parent: encodedData
+ });
+
+var Address = require('soop').load('bitcore/lib/Address',{
+  parent: versionedData
+});
+
+
 
 var TransactionDb = function() {
   TransactionDb.super(this, arguments);
@@ -463,6 +478,49 @@ TransactionDb.prototype.fromAddr = function(addr, cb) {
     });
 };
 
+TransactionDb.prototype._fromBuffer = function (buf) {
+    var buf2 = buffertools.reverse(buf);
+    return parseInt(buf2.toString('hex'), 16);
+};
+
+TransactionDb.prototype.getStandardizedTx = function (tx, time, isCoinBase) {
+  var self = this;
+  tx.txid = bitcoreUtil.formatHashFull(tx.getHash());
+  var ti=0;
+
+  tx.vin = tx.ins.map(function(txin) {
+    var ret = {n: ti++};
+    if (isCoinBase) {
+      ret.isCoinBase = true;
+    } else {
+      ret.txid = buffertools.reverse(new Buffer(txin.getOutpointHash())).toString('hex');
+      ret.vout = txin.getOutpointIndex();
+    }
+    return ret;
+  });
+
+  var to = 0;
+  tx.vout = tx.outs.map(function(txout) {
+    var val;
+    if (txout.s) {
+      var s = new Script(txout.s);
+      var addrs = new Address.fromScriptPubKey(s, config.network);
+      // support only for p2pubkey p2pubkeyhash and p2sh
+      if (addrs && addrs.length === 1) {
+        val = {addresses: [addrs[0].toString() ] };
+      }
+    }
+    return {
+      valueSat: self._fromBuffer(txout.v),
+      scriptPubKey: val,
+      n: to++,
+    };
+  });
+  tx.time = time;
+  return tx;
+};
+
+
 TransactionDb.prototype.fillScriptPubKey = function(txouts, cb) {
   var self=this;
   // Complete utxo info
@@ -513,7 +571,9 @@ TransactionDb.prototype._addScript = function(tx) {
   var relatedAddrs = [];
   var dbScript  = [];
   var ts        = tx.time;
-  var txid      = tx.txid;
+  var txid      = tx.txid || tx.hash;
+// var u=require('util');
+// console.log('[TransactionDb.js.518]', u.inspect(tx,{depth:10})); //TODO
   // Input Outpoints (mark them as spent)
   for(var ii in tx.vin) {
     var i = tx.vin[ii];
@@ -560,7 +620,6 @@ TransactionDb.prototype.add = function(tx, blockhash, cb) {
 
 TransactionDb.prototype._addManyFromObjs = function(txs, next) {
   var dbScript = [];
-
   for(var ii in txs){
     var s = this._addScript(txs[ii]);
     dbScript = dbScript.concat(s);
